move get cells map
