+++ python/helpers.py
@@ -3,6 +3,8 @@ import json
 import os
 import re
 
+######################################
+# Functions
 
 def path_to_list(path, ext=".root"):
     """
@@ -106,3 +108,37 @@ def simplify_dict(d):
     for k in old_keys:
         d[layer_number_from_string(k)] = d.pop(k)
     return d
+
+skip_pattern = r"(supportTube)|(cryo)"
+re_skip = re.compile(skip_pattern)
+
+def get_cells(detector, n_cells = 0):
+    """
+    Recursively count the number of cells in the detector sub-elements.
+    """
+    sub_detectors = detector.children()
+    if sub_detectors.size() == 0:
+        # print("Counting", detector.GetName())
+        # print("Counting", detector.id())
+        n_cells += 1
+    else:
+        for d in sub_detectors:
+            d_name = str(d[0])
+            if re_skip.match(d_name):
+                print("Skipping sub detector:", d_name)
+                continue
+            n_cells = get_cells(d[1], n_cells)
+    return n_cells
+
+######################################
+# Classes
+
+class GeoFile:
+    """
+    Class to handle geometry file information.
+    """
+    def __init__(self, path):
+        self.path = path                                                 # Full path to XML file
+        self.name = path.split("/")[-1].strip(".xml")                    # Get detector name and version
+        self.short = re.sub("_o[0-9]_v[0-9]{2}","",self.name)            # Get name only
+        self.version = re.search("o[0-9]_v[0-9]{2}",self.name).group(0)  # Get version only
